document everything, reorganize state machine example
