+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/examples/BasicStateMachine.java
@@ -6,6 +6,54 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import org.firstinspires.ftc.teamcode.subsystems.fsm.FiniteStateMachine;
 import org.firstinspires.ftc.teamcode.subsystems.fsm.FiniteStateMachineBuilder;
 
+import org.firstinspires.ftc.teamcode.subsystems.fsm.Edge;
+import org.firstinspires.ftc.teamcode.subsystems.fsm.State;
+
+/**
+ * A state representing moving forward.
+ */
+class ForwardState implements State {
+    @Override
+    public Edge[] getEdges() {
+        return new Edge[]{
+                new Edge("IdleState", (state) -> {
+                    // You can replace this with any logic that, when it returns true, will switch
+                    // the state machine over to IdleState.
+                    return false;
+                })
+        };
+    }
+
+    @Override
+    public String getName() {
+        return "ForwardState";
+    }
+}
+
+/**
+ * A state representing idelation.
+ */
+class IdleState implements State {
+    @Override
+    public Edge[] getEdges() {
+        return new Edge[]{
+                new Edge("ForwardState", (state) -> {
+                    // You can replace this with any logic that, when it returns true, will switch
+                    // the state machine over to ForwardState.
+                    return false;
+                })
+        };
+    }
+
+    @Override
+    public String getName() {
+        return "IdleState";
+    }
+}
+
+/**
+ * A basic state machine example, with no robot I/O.
+ */
 @TeleOp(name = "Basic State Machine", group = "Iterative OpMode")
 public class BasicStateMachine extends OpMode {
     /**
@@ -23,7 +71,9 @@ public class BasicStateMachine extends OpMode {
         telemetry.addData("Status", "Initializing");
 
         // Build the state machine.
-        FiniteStateMachineBuilder finiteStateMachineBuilder = new FiniteStateMachineBuilder();
+        FiniteStateMachineBuilder finiteStateMachineBuilder = new FiniteStateMachineBuilder()
+                .addState(new ForwardState())
+                .addState(new IdleState());
         finiteStateMachine = new FiniteStateMachine(finiteStateMachineBuilder);
 
         // Tell the driver the robot is ready.
