+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/examples/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains examples on how to use some subsystems.
+ */
+package org.firstinspires.ftc.teamcode.opmodes.examples;
