+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/fsm/MetaState.java
@@ -5,23 +5,29 @@ public class MetaState {
     private final State state;
     private StateState stateState;
 
-    /** Constructs a new MetaState. */
+    /** Constructs a new MetaState.
+     * @param state The underlying state.
+     * @param stateState The state of the state.
+     */
     public MetaState(State state, StateState stateState) {
         this.state = state;
         this.stateState = stateState;
     }
 
-    /** Gets the underlying state. */
+    /** Gets the underlying state.
+     * @return Ditto. */
     public State getState() {
         return state;
     }
 
-    /** Gets the state of the state. */
+    /** Gets the state of the state.
+     * @return Ditto. */
     public StateState getStateState() {
         return stateState;
     }
 
-    /** Sets the state of the state. */
+    /** Sets the state of the state.
+     * @param stateState Ditto. */
     public void setStateState(StateState stateState) {
         this.stateState = stateState;
     }
