+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/fsm/StateState.java
@@ -2,6 +2,9 @@ package org.firstinspires.ftc.teamcode.subsystems.fsm;
 
 /** The state of a State. */
 public enum StateState {
+    /** The state has yet to be run. */
     UNRAN,
+
+    /** The state has already been run, or is currently being run. */
     RAN,
 }
