+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/hardware/HardwareDeviceInitializer.java
@@ -4,5 +4,7 @@ import com.qualcomm.robotcore.hardware.HardwareDevice;
 
 /** Defines a callback that is run during a device initialization. */
 public interface HardwareDeviceInitializer {
+    /** Ditto.
+     * @param dev The hardware device to initialize. Must be casted. */
     void initialize(HardwareDevice dev) throws RuntimeException;
 }
