+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/hardware/TemplateDeviceInitializers.java
@@ -6,7 +6,9 @@ import com.qualcomm.robotcore.hardware.HardwareDevice;
 
 /** A collection of template device initializers. */
 public class TemplateDeviceInitializers {
-    /** Initializes a DcMotorEx to be reversed, brake on zero-power, and to use encoders. */
+    /** Initializes a DcMotorEx to be reversed, brake on zero-power, and to use encoders.
+     * @param dev The device to initialize.
+     */
     public static void motorReverseBrakeEncoder(HardwareDevice dev) {
         DcMotorEx motor = (DcMotorEx) dev;
         motor.setDirection(DcMotorEx.Direction.REVERSE);
@@ -14,7 +16,9 @@ public class TemplateDeviceInitializers {
         motor.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
     }
 
-    /** Initializes a DcMotorEx to go forward, brake on zero-power, and to use encoders. */
+    /** Initializes a DcMotorEx to go forward, brake on zero-power, and to use encoders.
+     * @param dev The device to initialize.
+     */
     public static void motorForwardBrakeEncoder(HardwareDevice dev) {
         DcMotorEx motor = (DcMotorEx) dev;
         motor.setDirection(DcMotorEx.Direction.FORWARD);
